// src/app/models/user.ts
export class User {
    id: string;
    username: string;
    firstName: string;
    lastName: string;
    email: boolean;
    password: string;
    confirmPassword: string;
    phone: string;
    crCode: string;

    roleKey: string;
    active: boolean;
    uid: string;
    nationality: string; 
}